Catch case where lexbox project not yet CRDT
Instead of downloading a CRDT project with no commits, which leads to
errors in FW Lite when it can't find a default writing system, we'll
check first whether the Lexbox project is a CRDT project, and refuse to
download it if it isn't. Currently this just displays an error
notification, but soon we'll display this in the dialog box instead.