Don't log newlines if we receive them
Copilot AI suggests that projectCode comes from user input and could
therefore contain newlines. That's not really likely given how the input
dialog is written, but it's theoretically possible. So to avoid newlines
in the middle of a log message, which could confuse some logging
software, we'll make this extra, not really needed, check.