Allow setting timeout value for idle service
Note that if useIdleService() is called in multiple places with
different values, the first one called will "win", and it's undefined
which one that will be. So it should be called with the same value
everywhere, or just change the DEFAULT_IDLE_TIMEOUT_MS value and call it
with no parameters everywhere.