Use index as #each key when uniqueness not guaranteed
In a few places, like when iterating over strings from a URL path, we
can't 100% guarantee that they will be unique, so it's better to use the
index as the key. This can result in unnecessary re-renderings, but this
is better than the result that would happen if a Svelte #each block
ended up with the same key pointing at multiple items in the #each. And
there won't be many re-renderings in most of these cases, as they are
not things that change frequently (if at all) in the UI.