Fix new eslint errors (#1844)
* Fix `svelte-check` warning about enableSourcemap

Sourcemaps are always on in Svelte 5, no need to enable them explicitly.

* Fix missing translation

This was caught by the svelte/no-useless-mustaches rule

* Fix remaining svelte/no-useless-mustaches errors

The two remaining svelte/no-useless-mustaches errors were auto-fixed by
the eslint rule.

* Fix eslint errors in Lexbox admin UI

FW Lite UI (frontend/viewer) will be next

* Fix eslint warnings in viewer

* Use index as #each key when uniqueness not guaranteed

In a few places, like when iterating over strings from a URL path, we
can't 100% guarantee that they will be unique, so it's better to use the
index as the key. This can result in unnecessary re-renderings, but this
is better than the result that would happen if a Svelte #each block
ended up with the same key pointing at multiple items in the #each. And
there won't be many re-renderings in most of these cases, as they are
not things that change frequently (if at all) in the UI.