Fix typing into empty rich-text fields
The prosemirror-view package, version 1.40.0, has an issue that causes
us to be unable to type into empty rich-text fields until the field
contents (which are nothing) are selected with Ctrl+A. Once a Select All
has happened, then the empty field can be typed into. Since our code
automatically does a Select All when focusing a field via keyboard
navigation such as Tab, this is only noticeable when clicking into a
field with the mouse and then trying to type.

Solving this issue by downgrading to prosemirror-view version 1.39.3.