Allow admins (site or org) to download projects by code (#1829)
* Backend now knows if user can download any project

Site admins and org admins can download arbitrary projects without
needing to be a member (projects owned by their org, or any project at
all in the case of site admins), so let's return a boolean flag for the
frontend to enable UI for that. The UI will still need to handle "Hey,
you're not actually authorized to download that project" because it will
be used by org admins, not only by site admins who have all access.

* Catch case where lexbox project not yet CRDT

Instead of downloading a CRDT project with no commits, which leads to
errors in FW Lite when it can't find a default writing system, we'll
check first whether the Lexbox project is a CRDT project, and refuse to
download it if it isn't. Currently this just displays an error
notification, but soon we'll display this in the dialog box instead.

* Much improved download-by-code dialog

Errors now show in dialog, dialog doesn't close until download
completed, dialog doesn't close if any errors.

* Also display error if user may not download project

Can happen if an org admin tries to download a project from another org

* Fix lint error

---------

Co-authored-by: Kevin Hahn <kevin_hahn@sil.org>