Create new MiniLcmApi wrapper for normalizing strings (#1873)
* Create new WrapWith method for wrapping MiniLcmApi

WrapWith wraps APIs in reverse order, last to first, so that the first
one in its parameter list is the first one to receive data from JS. That
way the ordering of factories in the array matches the order in which
data coming from JS will flow down through the wrappers into the API.

* Create normalization wrapper for MiniLcm search

* Add tests for NormalizationWrapper

Also hardcode normalization form to NFD, since we don't plan on ever
changing it.

* Fix BeaKona.AutoInterface invocation

Turns out AutoInterface needs BeaKona.MemberMatchTypes.Any for it to
actually use the implementations we were creating. So our wrappers
weren't actually doing anything yet, and we didn't know.